pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.0",
    "email-validator>=2.2.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "psycopg2-binary>=2.9.10",
    "python-dateutil>=2.9.0.post0",
    "pyyaml>=6.0.2",
    "sqlalchemy>=2.0.41",
    "watchdog>=6.0.0",
    "werkzeug>=3.1.3",
]
